refactor&feat: logout 로직 hook 으로 분리, 탈퇴성공시 logout 로직 실행 추가 #114
